feat: extract fetch mirrors timestamp
